+++ lib/src/platform/web/file_picker_web.dart
@@ -2,8 +2,12 @@ import 'dart:async';
 import 'dart:js_interop';
 import 'dart:typed_data';
 
-import 'package:file_picker_platform_interface/file_picker.dart';
-import 'package:file_picker_platform_interface/file_picker_platform_interface.dart';
+import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/api/file_picker_result.dart';
+import 'package:file_picker/src/api/platform_file.dart';
+import 'package:file_picker/src/api/android_saf_options.dart';
+import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
+import 'package:file_picker/src/utils/file_picker_utils.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:path/path.dart' as p;
 import 'package:web/web.dart';
@@ -62,11 +66,7 @@ class FilePickerWeb extends FilePickerPlatform {
     int compressionQuality = 0,
     AndroidSAFOptions? androidSafOptions,
   }) async {
-    if (type != FileType.custom && (allowedExtensions?.isNotEmpty ?? false)) {
-      throw Exception(
-        'You are setting a type [$type]. Custom extension filters are only allowed with FileType.custom, please change it or remove filters.',
-      );
-    }
+    FilePickerUtils.validateAllowedExtensions(type, allowedExtensions);
 
     Completer<List<PlatformFile>?>? filesCompleter =
         Completer<List<PlatformFile>?>();